src/a2a_mcp/agents/adk_examples/linkedin_carousel_agent/requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0